Feat : 실제 투표 생성 시 블록체인 네트워크 자동 배포 요청
블록체인 네트워크 자동 배포가 약 10초가 소요되어 @Async을 사용하여 비동기 처리하였다.

+++ Backend/src/main/java/com/ssafy/pickit/domain/voteSession/application/service/VoteSessionService.java
@@ -3,6 +3,7 @@ package com.ssafy.pickit.domain.voteSession.application.service;
 import java.time.LocalDateTime;
 import java.util.Comparator;
 import java.util.List;
+import java.util.concurrent.CompletableFuture;
 
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
@@ -24,14 +25,24 @@ import lombok.RequiredArgsConstructor;
 public class VoteSessionService {
 	private final VoteSessionRepository voteSessionRepository;
 	private final TempVoteSessionService tempVoteSessionService;
+	private final VoteSessionDeployService voteSessionDeployService;
 
 	@Transactional
-	public VoteSession create(String id, String contractAddress) {
+	public void create(String id) {
 		TempVoteSession tempVoteSession = tempVoteSessionService.findById(id);
-		VoteSession voteSession = buildVoteSession(tempVoteSession, contractAddress, mapToCandidates(tempVoteSession));
-		tempVoteSessionService.delete(id);
+		CompletableFuture<String> completableFuture = voteSessionDeployService.deploy(tempVoteSession.getCandidates());
 
-		return voteSessionRepository.save(voteSession);
+		completableFuture.thenApply(contractAddress -> {
+			if (contractAddress != null) {
+				VoteSession voteSession = buildVoteSession(tempVoteSession, contractAddress,
+					mapToCandidates(tempVoteSession));
+				tempVoteSessionService.delete(id);
+
+				return voteSessionRepository.save(voteSession);
+			} else {
+				throw new RuntimeException("VoteSession deploy failed.");
+			}
+		});
 	}
 
 	public List<VoteSessionListResponse> findAllByOngoing() {
